% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{er_add_ffmpeg_path}
\alias{er_add_ffmpeg_path}
\title{Add the path to ffmpeg to the system path}
\usage{
er_add_ffmpeg_path()
}
\value{
\code{TRUE} on success, invisibly
}
\description{
If \code{ffmpeg} can already be seen by the system, the system path will not be changed. Otherwise, the system path will be prepended with the ffmpeg path.
}
