% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_audio_track}
\alias{er_audio_track}
\title{Audio track}
\usage{
er_audio_track(path, cut_from, cut_to, start, mix_volume)
}
\arguments{
\item{path}{string: path to audio file}

\item{cut_from}{numeric: time value to cut from, in seconds}

\item{cut_to}{numeric: time value to cut to, in seconds}

\item{start}{numeric: how many seconds into the video should this audio track start? (default = 0)}

\item{mix_volume}{numeric: relative volume when mixing this video's audio track with others (default = 1)}
}
\value{
A list of class \code{er_audio_track}
}
\description{
The difference between \code{er_audio_track} and \code{\link[=er_layer_audio]{er_layer_audio()}} is that an \code{er_audio_track} will continue to play across multiple clips and can start and stop whenever needed.
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_header]{er_header()}}, \code{\link[=er_spec]{er_spec()}}
}
