% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_clip_pause}
\alias{er_clip_pause}
\alias{er_clip_title2}
\title{Convenience functions for clips}
\usage{
er_clip_pause(duration, ...)

er_clip_title(duration, ...)

er_clip_title_background(duration, ...)

er_clip_image(duration, ...)

er_clip_image_overlay(duration, ...)

er_clip_subtitle(duration, ...)

er_clip_news_title(duration, ...)

er_clip_slide_in_text(duration, ...)

er_clip_fill_color(duration, ...)

er_clip_radial_gradient(duration, ...)

er_clip_linear_gradient(duration, ...)

er_clip_rainbow_colors(duration, ...)

er_clip_video(duration, ...)

er_clip_audio(duration, ...)

er_clip_title2(duration, ...)
}
\arguments{
\item{duration}{numeric: duration in seconds}

\item{...}{: as for the corresponding \verb{er_layer_*} function}
}
\value{
A list of class \code{er_clip}
}
\description{
Each of the \verb{er_clip_*} functions is a wrapper around the corresponding \verb{er_layer_*} function, creating an \code{er_clip} object with the structure \code{er_clip(duration, transition, layer)}, where:
\itemize{
\item \code{transition} is optional and only added if it appears in the \code{...} parameters
\item \code{layer} is created via the \verb{er_layer_*} function, passing the appropriate parameters from \code{...}
}
}
\details{
\code{er_clip_title2} is the same as \code{er_clip_title} but automatically inserts an \code{er_layer_fill_color} layer underneath the title, to fix font rendering issues (see \url{https://github.com/mifi/editly/issues/84})
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_clip]{er_clip()}}
}
\keyword{internal} \alias{er_clip_pause}
\keyword{internal} \alias{er_clip_title}
\keyword{internal} \alias{er_clip_title_background}
\keyword{internal} \alias{er_clip_image}
\keyword{internal} \alias{er_clip_image_overlay}
\keyword{internal} \alias{er_clip_subtitle}
\keyword{internal} \alias{er_clip_news_title}
\keyword{internal} \alias{er_clip_slide_in_text}
\keyword{internal} \alias{er_clip_fill_color}
\keyword{internal} \alias{er_clip_radial_gradient}
\keyword{internal} \alias{er_clip_linear_gradient}
\keyword{internal} \alias{er_clip_rainbow_colors}
\keyword{internal} \alias{er_clip_video}
\keyword{internal} \alias{er_clip_audio}
