% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{er_exec_wait}
\alias{er_exec_wait}
\title{Execute editly}
\usage{
er_exec_wait(json, fast = FALSE, ...)
}
\arguments{
\item{json}{string: path to json file or json string, as returned by \code{\link[=er_spec]{er_spec()}}}

\item{fast}{logical: if \code{TRUE}, run editly in 'fast' (low quality/preview) mode}

\item{...}{: if a json file is not being used, you can provide command line arguments as individual parameters}
}
\value{
0 for success (invisibly)
}
\description{
Execute editly
}
\examples{
\dontrun{
  er_exec_wait()
}

}
\references{
\url{https://github.com/mifi/editly/}
}
