% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{er_ffmpeg_exe}
\alias{er_ffmpeg_exe}
\title{Path to the ffmpeg executable}
\usage{
er_ffmpeg_exe()
}
\value{
The path to the executable, or \code{NULL} if not found
}
\description{
Path to the ffmpeg executable
}
\examples{
er_ffmpeg_exe()

}
\seealso{
\code{\link[=er_install_ffmpeg]{er_install_ffmpeg()}}
}
