% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{er_install_editly}
\alias{er_install_editly}
\title{Install the editly node package}
\usage{
er_install_editly(force = FALSE, dependencies = TRUE)
}
\arguments{
\item{force}{logical: if \code{TRUE}, force reinstallation of editly if it is already installed}

\item{dependencies}{logical: if \code{TRUE}, the system binaries \code{node} (via \code{\link[noder:nr_install_node]{noder::nr_install_node()}}) and \code{ffmpeg} (via \code{\link[=er_install_ffmpeg]{er_install_ffmpeg()}}) will also be installed if not found on your system}
}
\value{
The path to the installation (invisibly)
}
\description{
Editry requires that \code{editly} (the underlying nodeJS package) be available on your system, along with \code{node} and \code{ffmpeg}. These can either be installed directly on your system, or in a docker image. \code{er_install_editly} will attempt to install \code{node}, \code{ffmpeg}, and \code{editly} directly on your system. See the package overview at \url{https://scienceuntangled.github.io/editry} for docker-based installation.
}
\references{
\url{https://github.com/scienceuntangled/editly/}
}
