% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{er_install_ffmpeg}
\alias{er_install_ffmpeg}
\title{Install ffmpeg}
\usage{
er_install_ffmpeg(force = FALSE, check_hash = TRUE)
}
\arguments{
\item{force}{logical: force reinstallation if ffmpeg already exists}

\item{check_hash}{logical: don't check the hash of the downloaded file}
}
\value{
the path to the installed executable
}
\description{
This is a helper function to install ffmpeg. Currently it only works on Windows platforms. The ffmpeg bundle will be downloaded from \url{https://www.gyan.dev/ffmpeg/builds/} and saved to your user appdata directory.
}
\examples{
\dontrun{
  er_install_ffmpeg()
}

}
\references{
\url{https://www.gyan.dev/ffmpeg/builds/}
}
