% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_layer}
\alias{er_layer}
\alias{er_clip}
\alias{er_defaults}
\alias{er_header}
\title{Create various sections of an editly spec}
\usage{
er_layer(type, ...)

er_clip(duration, transition, layers)

er_defaults(transition, layer, layer_type)

er_header(
  out_path = tempfile(fileext = ".mp4"),
  fps,
  width,
  height,
  allow_remote_requests,
  audio_file_path,
  loop_audio,
  keep_source_audio,
  clips_audio_volume,
  output_volume,
  audio_norm.enable,
  audio_norm.gauss_size,
  audio_norm.max_gain,
  defaults,
  custom_output_args,
  logo_path,
  logo_width,
  logo_height,
  logo_x,
  logo_y
)
}
\arguments{
\item{type}{string: the layer type (e.g. "title", "image", "video")}

\item{...}{: other parameters}

\item{duration}{numeric: duration in seconds}

\item{transition}{list: an \code{\link[=er_transition]{er_transition()}} object}

\item{layers}{list: a list of \code{er_layer} objects}

\item{layer}{list: default layer spec}

\item{layer_type}{list: default layerType spec}

\item{out_path}{string: the path of the video file to create (the file extension can be a video file extension or ".gif")}

\item{fps}{numeric: frames per second (default = 25 or FPS of first video)}

\item{width}{integer: video width (default = 640)}

\item{height}{integer: auto based on width and aspect ratio of first video}

\item{allow_remote_requests}{logical: allow remote URLs as paths? (default = \code{FALSE})}

\item{audio_file_path}{string: set an audio track for the whole video}

\item{loop_audio}{logical: loop the audio track if it is shorter than video? (default = \code{FALSE})}

\item{keep_source_audio}{logical: keep source audio from clips? (default = \code{FALSE})}

\item{clips_audio_volume}{numeric: volume of audio from clips relative to \code{\link[=er_audio_track]{er_audio_track()}}s (default = 1)}

\item{output_volume}{numeric or string: adjust output volume (final stage). e.g. 0.5 or "10dB" (default = 1)}

\item{audio_norm.enable}{logical: enable audio normalization? (default = \code{FALSE})}

\item{audio_norm.gauss_size}{numeric: audio normalization gauss size (default = 5)}

\item{audio_norm.max_gain}{numeric: audio normalization max gain (default = 30)}

\item{defaults}{er_defaults: as returned by \code{\link[=er_defaults]{er_defaults()}}}

\item{custom_output_args}{character: vector of custom output codec/format arguments for ffmpeg}

\item{logo_path}{string: path to logo file. This must be a local file, even if \code{allow_remote_requests} is \code{TRUE}. \code{logo_path} will be ignored if \code{out_path} is a gif}

\item{logo_width}{numeric: width of logo relative to screen width (0 to 1, default = 0.2). If only one of logo height or width is specified, the logo aspect ratio will be retained}

\item{logo_height}{numeric: height of logo relative to screen width (0 to 1, default = 0.1). If only one of logo height or width is specified, the logo aspect ratio will be retained}

\item{logo_x}{numeric: logo left-hand-edge position relative to left of screen (0 = left to 1 = right, default = 0.78). The logo is referenced relative to its left-hand edge, so the default of 0.78 and width 0.2 puts the right-hand edge at 0.98}

\item{logo_y}{numeric: logo bottom-edge position relative to top of screen (0 = top to 1 = bottom, default = 0.98)}
}
\value{
A list with class \code{er_clip}, \code{er_layer}, etc
}
\description{
Create various sections of an editly spec
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_spec]{er_spec()}}
}
