% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_layer_pause}
\alias{er_layer_pause}
\alias{er_layer_image}
\alias{er_layer_image_overlay}
\alias{er_layer_title}
\alias{er_layer_subtitle}
\alias{er_layer_title_background}
\alias{er_layer_news_title}
\alias{er_layer_slide_in_text}
\alias{er_layer_fill_color}
\alias{er_layer_radial_gradient}
\alias{er_layer_linear_gradient}
\alias{er_layer_rainbow_colors}
\alias{er_layer_video}
\alias{er_layer_audio}
\alias{er_layer_detached_audio}
\title{Standard layer types}
\usage{
er_layer_pause(color)

er_layer_image(path, resize_mode, zoom_direction, zoom_amount)

er_layer_image_overlay(
  path,
  position,
  width,
  height,
  zoom_direction,
  zoom_amount
)

er_layer_title(
  text,
  font_path,
  text_color,
  position,
  zoom_direction,
  zoom_amount
)

er_layer_subtitle(text, font_path, text_color)

er_layer_title_background(text, font_path, text_color, background)

er_layer_news_title(text, font_path, text_color, background_color, position)

er_layer_slide_in_text(
  text,
  font_path,
  font_size,
  char_spacing,
  color,
  position
)

er_layer_fill_color(color)

er_layer_radial_gradient(colors)

er_layer_linear_gradient(colors)

er_layer_rainbow_colors()

er_layer_video(
  path,
  resize_mode,
  cut_from,
  cut_to,
  width,
  height,
  left,
  top,
  origin_x,
  origin_y,
  mix_volume
)

er_layer_audio(path, cut_from, cut_to, mix_volume)

er_layer_detached_audio(path, cut_from, cut_to, start, mix_volume)
}
\arguments{
\item{color}{string: slide color (default = random)}

\item{path}{string: path to image or video file}

\item{resize_mode}{: see Details}

\item{zoom_direction}{string: "in", out", or "null" to disable}

\item{zoom_amount}{numeric: amount to zoom (default = 0.1)}

\item{position}{string or list: see Details}

\item{width}{numeric: width (from 0 to 1) where 1 is screen width}

\item{height}{numeric: height (from 0 to 1) where 1 is screen height}

\item{text}{string: text to show}

\item{font_path}{string: path to font}

\item{text_color}{string: text color}

\item{background}{er_layer: as returned by \code{\link[=er_layer_radial_gradient]{er_layer_radial_gradient()}}, \code{\link[=er_layer_linear_gradient]{er_layer_linear_gradient()}}, or \code{\link[=er_layer_fill_color]{er_layer_fill_color()}}}

\item{background_color}{string: background color}

\item{font_size}{numeric: font size}

\item{char_spacing}{numeric: character spacing}

\item{colors}{character: vector of two colors (default = random)}

\item{cut_from}{numeric: time value to cut from, in seconds}

\item{cut_to}{numeric: time value to cut to, in seconds}

\item{left}{numeric: X-position relative to screen width (0 to 1, default = 0)}

\item{top}{numeric: Y-position relative to screen height (0 to 1, default = 0)}

\item{origin_x}{numeric: X anchor ("left" or "right", default = "left")}

\item{origin_y}{numeric: Y anchor ("top" or "bottom", default = "top")}

\item{mix_volume}{numeric: relative volume when mixing this video's audio track with others (default = 1)}

\item{start}{numeric: how many seconds into this clip should this audio track start? (default = 0)}
}
\value{
A list with class \code{er_layer}
}
\description{
position can be one of either:
\itemize{
\item top, bottom center, top-left, top-right, center-left, center-right, bottom-left, bottom-right
\item a list object {x, y, originX = 'left', originY = 'top' }, where { x: 0, y: 0 } is the upper left corner of the screen, and { x: 1, y: 1 } is the lower right corner, x is relative to video width, y to video height. originX and originY are optional, and specify the position's origin (anchor position) of the object.
}
}
\details{
Audio layers will be mixed together. If cutFrom/cutTo is set, the resulting segment (cutTo-cutFrom) will be slowed/sped-up to fit clip.duration. The slow down/speed-up operation is limited to values between 0.5x and 100x.
\code{er_layer_detached_audio} is a special case of \code{audio_tracks} that makes it easier to start the audio relative to clips start times without having to calculate global start times. detached audio has the exact same properties as \code{audio_tracks}, except start time is relative to the clip's start.

For video layers, if parent clip.duration is specified, the video will be slowed/sped-up to match clip.duration. If cutFrom/cutTo is set, the resulting segment (cutTo-cutFrom) will be slowed/sped-up to fit clip.duration. If the layer has audio, it will be kept (and mixed with other audio layers if present.)

resize_mode - How to fit image to screen. Can be one of:
\itemize{
\item "contain" - All the video will be contained within the frame and letterboxed
\item "contain-blur" - Like contain, but with a blurred copy as the letterbox
\item "cover" - Video be cropped to cover the whole screen (aspect ratio preserved)
\item "stretch" - Video will be stretched to cover the whole screen (aspect ratio ignored).
Default is "contain-blur".
}
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_layer]{er_layer()}}
}
