% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{er_theme_default}
\alias{er_theme_default}
\title{Default editry theme}
\usage{
er_theme_default()
}
\value{
An object of class \code{er_defaults}
}
\description{
Themes in editry are a set of defaults that can be supplied to the \code{\link[=er_header]{er_header()}} or \code{\link[=er_spec]{er_spec()}} functions.
}
\details{
The default theme currently uses:
\itemize{
\item white text, with 'PatuaOne' font face
\item black slide backgrounds for plain slides, and grey gradients for radial or linear gradient backgrounds
\item red background for news-title layers
\item resize_mode "contain" for images
}
}
\seealso{
\code{\link[=er_defaults]{er_defaults()}}, \code{\link[=er_spec]{er_spec()}}
}
