% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_to_json}
\alias{er_to_json}
\title{Convert editly spec to JSON string}
\usage{
er_to_json(spec)
}
\arguments{
\item{spec}{string: a JSON string as returned by \code{\link[=er_spec]{er_spec()}}}
}
\value{
A JSON string
}
\description{
Convert editly spec to JSON string
}
