% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_transition}
\alias{er_transition}
\title{Clip transitions}
\usage{
er_transition(name, duration, audio_out_curve, audio_in_curve)
}
\arguments{
\item{name}{string: can be any of the transitions listed in \url{https://gl-transitions.com/gallery}, or "directional-left", "directional-right", "directional-up", "directional-down", "random", " dummy"}

\item{duration}{numeric: duration in seconds}

\item{audio_out_curve}{string: fade out curve in audio cross fades (default = "tri")}

\item{audio_in_curve}{string: fade in curve in audio cross fades (default = "tri")}
}
\value{
A list with class \code{er_transition}
}
\description{
Clip transitions
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_clip]{er_clip()}}
}
