% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_clip_pause}
\alias{er_clip_pause}
\alias{er_clip_pausetitle}
\alias{er_clip_title}
\alias{er_clip_title_background}
\alias{er_clip_image}
\title{Convenience functions for clips}
\usage{
er_clip_pause(duration, ...)

er_clip_pausetitle(duration, ...)

er_clip_title(duration, ...)

er_clip_title_background(duration, ...)

er_clip_image(duration, ...)
}
\arguments{
\item{duration}{numeric: duration in seconds}

\item{...}{: as for the corresponding \verb{er_layer_*} function}
}
\value{
A list of class \code{er_clip}
}
\description{
Convenience functions for clips
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_clip]{er_clip()}}
}
