% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{er_exec_wait}
\alias{er_exec_wait}
\title{Execute editly}
\usage{
er_exec_wait(spec, fast = FALSE, ...)
}
\arguments{
\item{spec}{er_spec or string: \code{er_spec} object, or a string giving the path to a JSON file or a JSON string, as returned by \code{\link[=er_spec]{er_spec()}}}

\item{fast}{logical: if \code{TRUE}, run editly in 'fast' (low quality/preview) mode}

\item{...}{: if a JSON file is not being used, you can provide command line arguments as individual parameters}
}
\value{
0 for success (invisibly)
}
\description{
Execute editly
}
\examples{
\dontrun{
  clips <- list(er_clip_title2(duration = 3, text = "Title slide"),
                er_clip_image(duration = 3, path = "https://jeroen.github.io/images/Rlogo.png"))
  outfile <- tempfile(fileext = ".mp4")
  my_spec <- er_spec(clips = clips , out_path = outfile, allow_remote_requests = TRUE)
  er_exec_wait(my_spec)
  if (interactive()) browseURL(outfile)
}

}
\references{
\url{https://github.com/mifi/editly/}
}
