% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{er_install_editly}
\alias{er_install_editly}
\title{Install the editly node package}
\usage{
er_install_editly(force = FALSE)
}
\arguments{
\item{force}{logical: if \code{TRUE}, force reinstallation of editly if it is already installed}
}
\value{
The path to the installation (invisibly)
}
\description{
Install the editly node package
}
\references{
\url{https://github.com/mifi/editly/}
}
