% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{er_install_editly}
\alias{er_install_editly}
\title{Install the editly node package}
\usage{
er_install_editly()
}
\value{
The path to the installation
}
\description{
Install the editly node package
}
\references{
\url{https://github.com/mifi/editly/}
}
