% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{er_install_ffmpeg}
\alias{er_install_ffmpeg}
\title{Install ffmpeg}
\usage{
er_install_ffmpeg(force = FALSE, bits, check_hash = TRUE)
}
\arguments{
\item{force}{logical: force reinstallation if ffmpeg already exists}

\item{bits}{integer: 32 or 64, for 32- or 64-bit install. If missing or \code{NULL}, will be guessed based on \code{.Machine$sizeof.pointer}. Note that only 64-bit is supported on Windows}

\item{check_hash}{logical: don't check the hash of the downloaded file. Ignored on windows}
}
\value{
the path to the installed executable
}
\description{
This is a helper function to install ffmpeg. Currently it only works on Windows and Linux platforms. The ffmpeg bundle will be downloaded from \url{https://github.com/BtbN/FFmpeg-Builds/releases/latest} (Windows) or \url{https://johnvansickle.com/ffmpeg/} (Linux) and saved to your user appdata directory.
}
\examples{
\dontrun{
  er_install_ffmpeg()
}

}
\references{
\url{https://github.com/BtbN/FFmpeg-Builds/releases/latest} \url{https://johnvansickle.com/ffmpeg/}
}
