% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_layer}
\alias{er_layer}
\alias{er_clip}
\alias{er_defaults}
\alias{er_header}
\title{Create various sections of an editly spec}
\usage{
er_layer(type, ...)

er_clip(duration, layers)

er_defaults(transition, layer, layer_type)

er_header(fps = 30, out_path, defaults)
}
\arguments{
\item{type}{string: the layer type (e.g. "title" or "image")}

\item{...}{: other parameters}

\item{duration}{numeric: clip duration in seconds}

\item{layers}{list: a list of \code{er_layer} objects}

\item{transition}{list: default transition spec}

\item{layer}{list: default layer spec}

\item{layer_type}{list: default layerType spec}

\item{fps}{numeric: frames per second}

\item{out_path}{string: the path of the video file to create}

\item{defaults}{er_defaults: as returned by \code{\link[=er_defaults]{er_defaults()}}}
}
\value{
A list with class \code{er_clip}, \code{er_layer}, etc
}
\description{
Create various sections of an editly spec
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_spec]{er_spec()}}
}
