% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_spec}
\alias{er_spec}
\title{Create the json string ready for compiling to video by editly}
\usage{
er_spec(header, clips, audio_tracks)
}
\arguments{
\item{header}{er_header: as returned by \code{\link[=er_header]{er_header()}}, or a list containing the header, clips, and (optionally) audio tracks, equivalent to \code{list(header, clips = list(clips), audioTracks = list(audio_tracks))}}

\item{clips}{list: a list of \code{er_clip} objects as returned by \code{\link[=er_clip]{er_clip()}}}

\item{audio_tracks}{list: a list of \code{er_audio_track} objects, as returned by \code{\link[=er_audio_track]{er_audio_track()}}}
}
\value{
A json string
}
\description{
Create the json string ready for compiling to video by editly
}
\references{
\url{https://github.com/mifi/editly/}
}
\seealso{
\code{\link[=er_layer]{er_layer()}}, \code{\link[=er_exec_wait]{er_exec_wait()}}
}
