% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editry.R
\name{er_theme_default}
\alias{er_theme_default}
\title{Editry themes}
\usage{
er_theme_default()
}
\value{
An object of class \code{er_defaults}
}
\description{
Editry themes
}
\seealso{
\code{\link[=er_defaults]{er_defaults()}}, \code{\link[=er_spec]{er_spec()}}
}
